pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "material-color-utilities-py"
version = "0.1.3"
description = "Python port of material-color-utilities used for Material You colors"
authors = [
    {name = "Deathemonic", email = "deathemonic@gmail.com"},
    {name = "Avanish Subbiah", email = "avanishsubbiah1@gmail.com"},
    {name = "0xMRTT"},
    {name = "David Lapshin"}
]
maintainers = [{name = "Deathemonic", email = "deathemonic@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache 2.0"}
classifiers = [
    "Environment :: X11 Applications",
    "License :: OSI Approved :: Apache 2.0 License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3"
]
dependencies = [
    "regex",
    "numpy",
    "numba; python_version<'3.11'",
    "Pillow >= 9.2.0"
]